std.child_process: enable non-standard streams
The main problem of process spawning methods on Unix (clone within
fork() or posix_spawn) is 1. CLOEXEC takes not immediately effect
with the clone(), but after the setup code for the child process was
run and a exec* function is executed.  This file descriptor can be
leaked to another process, which is especially problematic for
long-running processes or if no exec* function is called.

Additionally, to my knowledge all Unix Kernels have no way to 2. provide
the Kernel with the list of file descriptors which the child process
should inherit to prevent other threads from leaking a file descriptor
to another spawned process while it is inheritable.

3. Windows has an api to use fix problem 1+2, but it does not prevent
external code not to leak handles (file descriptors on Windows) via
different Kernel call.

=> This means that we can and should provide the user with the tools
   to handle it on Windows.

On Unixes can only
- 1. use slow user-space barriers to synchronize all potentially
  spawning threads,
- 2. accept potential leaks (and reduce time it remains open with CLOEXEC)
- 3. manually track file descriptor numbers along processes:
  On interaction with unknown external programs, one typically closes
  all non-standard handles before clone()/fork() and reopens them
  afterwards.

With at least Windows providing a nice way to prevent leakage into other
child processes (if all threads use the correct process spawn call):
- 1. make things uniform to make doing the correct thing on Windows simple
- 2. use allocations = posix spawn like api to reduce possible time to leak
     on Linux
- 3. tests as example code for the 3 ways to setup non-standard pipes
     (1. stdin, 2. pipe-stdin, 3. environment variables)
- 4. direct ways for the user to control the posix_spawn attributes and
     actions except for common default options and to "just spawn a process"
     to reduce time for leaking and do other common child process setup steps.